+++ src/core/server/api/consts/constConfig.ts
@@ -1,0 +1,5 @@
+import { PlayerConfig } from '@AthenaServer/systems/playerConfig';
+
+export const configConst = {
+    player: PlayerConfig,
+};