+++ src/core/server/imports/systems.ts
@@ -21,4 +21,5 @@ import '../systems/vehicle';
 import '../systems/defaults/death';
 import '../systems/defaults/inventory';
 import '../systems/defaults/time';
+import '../systems/defaults/weaponItems';
 import '../systems/defaults/weather';