+++ src/core/server/systems/defaults/inventory.ts
@@ -1,9 +1,9 @@
 import * as alt from 'alt-server';
 import { ATHENA_EVENTS_PLAYER } from '@AthenaShared/enums/athenaEvents';
-import { PLAYER_LOCAL_META } from '@AthenaShared/enums/playerSynced';
 import { StoredItem } from '@AthenaShared/interfaces/item';
 import { Athena } from '@AthenaServer/api/athena';
 import { PluginSystem } from '../plugins';
+import { SYSTEM_EVENTS } from '@AthenaShared/enums/system';
 
 let enabled = true;
 
@@ -14,13 +14,20 @@ const Internal = {
      *
      * @param {alt.Player} player
      */
-    sync(player: alt.Player) {
+    sync(player: alt.Player, isCharacterSelect = true) {
         if (!enabled) {
             return;
         }
 
         if (!player || !player.valid) {
             return;
+        }
+
+        if (isCharacterSelect) {
+            const config = Athena.systems.itemManager.config.get();
+            Athena.config.player.set(player, 'inventory-size', config.inventory.size);
+            Athena.config.player.set(player, 'inventory-weight-enabled', config.weight.enabled);
+            Athena.config.player.set(player, 'inventory-max-weight', config.weight.player);
         }
 
         const data = Athena.document.character.get(player);
@@ -34,8 +41,11 @@ const Internal = {
         const fullInventory = Athena.systems.itemManager.inventory.convertFromStored(inventory);
         const fullToolbar = Athena.systems.itemManager.inventory.convertFromStored(toolbar);
 
-        player.setLocalMeta(PLAYER_LOCAL_META.INVENTORY, fullInventory);
-        player.setLocalMeta(PLAYER_LOCAL_META.TOOLBAR, fullToolbar);
+        let totalWeight = 0;
+        totalWeight += Athena.systems.itemManager.inventory.getWeight(inventory);
+        totalWeight += Athena.systems.itemManager.inventory.getWeight(toolbar);
+
+        player.emit(SYSTEM_EVENTS.PLAYER_EMIT_INVENTORY_SYNC, fullInventory, fullToolbar, totalWeight);
     },
     /**
      * Automatically synchronizes the inventory for a player when data is changed.
@@ -48,7 +58,7 @@ const Internal = {
             return;
         }
 
-        Internal.sync(player);
+        Internal.sync(player, false);
     },
     /**
      * Automatically synchronizes the toolbar for a player when data is changed.
@@ -61,7 +71,7 @@ const Internal = {
             return;
         }
 
-        Internal.sync(player);
+        Internal.sync(player, false);
     },
     /**
      * Initializes the initial events for handling document changes.