+++ src/core/shared/enums/globalSynced.ts
@@ -1,0 +1,3 @@
+export enum GLOBAL_SYNCED {
+    INVENTORY_WEIGHT_ENABLED = 'inventory-weight-enabled',
+}