+++ src/core/shared/enums/playerConfigKeys.ts
@@ -1,0 +1,2 @@
+export type PlayerConfigKeys = //
+    'inventory-size' | 'inventory-weight-enabled' | 'inventory-max-weight';